feat: 카테고리 리스트 스크롤 생성
