+++ src/pages/ProductRegisterPage.js
@@ -1,11 +1,13 @@
 import ProductOption from "components/ProductOption";
 import React from "react";
 import styled from "styled-components";
+import CategoryPage from "./CategoryPage";
 
 function ProductRegisterPage() {
   return (
     <ProductRegisterPageContainer>
       <ProductOption />
+      <CategoryPage />
     </ProductRegisterPageContainer>
   );
 }
